Refactored authentification component
