+++ angular-client/src/app/modules/authentification/account/account.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthService } from '../../shared/services/auth.service';
+import { AuthService } from '@services';
 
 @Component({
   selector: 'app-account',
