+++ angular-client/src/app/modules/authentification/auth.routes.ts
@@ -1,6 +1,6 @@
 import {Routes} from '@angular/router';
-import {SignupComponent} from './signup.component';
-import {SigninComponent} from './signin.component';
+import {SignupComponent} from './signup/signup.component';
+import {SigninComponent} from './signin/signin.component';
 
 export const AUTH_ROUTES: Routes = [
   { path: '', redirectTo: 'signup', pathMatch: 'full' },
