+++ angular-client/src/app/modules/authentification/authentification-guard.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@angular/core';
 import { CanActivate } from '@angular/router';
-import { AuthService } from '../../shared/services/auth.service';
+import { AuthService } from '@services';
+
 
 @Injectable()
-export class CanActivateViaAuthGuard implements CanActivate {
+export class CanActivateAuthentificationGuard implements CanActivate {
 
   constructor(private authService: AuthService) {}
