+++ angular-client/src/app/modules/authentification/index.ts
@@ -1,0 +1,3 @@
+export { AccountComponent } from './account/account.component';
+export { SigninComponent } from './signin/signin.component';
+export { SignupComponent } from './signup/signup.component';
