+++ angular-client/src/app/modules/authentification/signin/signin.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from "@angular/core"
 import {FormControl, FormGroup, Validators} from "@angular/forms";
-import {User} from "../../shared/models/user.model";
-import {AuthService} from "../../shared/services/auth.service";
+import {User} from "../../../shared/models/user.model";
+import {AuthService} from "../../../shared/services/auth.service";
 import {Router} from "@angular/router";
 
 @Component({
