+++ angular-client/src/app/shared/services/index.ts
@@ -1,1 +1,2 @@
 export { ErrorService} from './error.service';
+export { AuthService} from './auth.service';
